app/requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==4.2.16
djangorestframework==3.15.2
drf-spectacular==0.27.1
environ==1.0
exceptiongroup==1.2.2
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.1
pillow==11.0.0
pluggy==1.5.0
pytest==8.3.3
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2
uritemplate==4.1.1
